backend/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Redis
redis==5.0.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# OCR
easyocr==1.7.1
opencv-python==4.8.1.78
Pillow==10.1.0

# LLM APIs
anthropic==0.7.7
openai==1.3.7

# Image processing
numpy==1.26.2

# Configuration
pydantic==2.5.2
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
